Migration Resolve and optimization

+++ app/0مي/2022_10_31_165427_add_rename_delivery_charge_column_to_stores_table.php
@@ -14,8 +14,16 @@ class AddRenameDeliveryChargeColumnToStoresTable extends Migration
     public function up()
     {
         Schema::table('stores', function (Blueprint $table) {
-            $table->renameColumn('delivery_charge', 'minimum_shipping_charge');
-            $table->double('per_km_shipping_charge',16, 3, true)->default('0');
+            if (
+                Schema::hasColumn('stores', 'minimum_shipping_charge')
+                && !Schema::hasColumn('stores', 'delivery_charge')
+            ) {
+                $table->renameColumn('minimum_shipping_charge', 'delivery_charge');
+            }
+
+            if (Schema::hasColumn('stores', 'per_km_shipping_charge')) {
+                $table->dropColumn('per_km_shipping_charge');
+            }
         });
     }
 
@@ -27,8 +35,16 @@ class AddRenameDeliveryChargeColumnToStoresTable extends Migration
     public function down()
     {
         Schema::table('stores', function (Blueprint $table) {
-            $table->renameColumn('minimum_shipping_charge', 'delivery_charge');
-            $table->dropColumn('per_km_shipping_charge');
+            if (
+                Schema::hasColumn('stores', 'minimum_shipping_charge')
+                && !Schema::hasColumn('stores', 'delivery_charge')
+            ) {
+                $table->renameColumn('minimum_shipping_charge', 'delivery_charge');
+            }
+
+            if (Schema::hasColumn('stores', 'per_km_shipping_charge')) {
+                $table->dropColumn('per_km_shipping_charge');
+            }
         });
     }
 }
